disabled and tooltip for useDetectionMode
